# data-reader-api-server/src/swagger.yml
openapi: 3.0.0
info:
    title: Data Reader API
    description: API documentation for Distributed Data Repository's Data Reader API calls. Mainly purposed for reading. Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
    version: 1.0.0
servers:
    - url: http://api.example.com/v1
      description: Optional server description, e.g. Main (production) server
    - url: http://staging-api.example.com
      description: Optional server description, e.g. Internal staging server for testing
components:
    securitySchemes:
        oauth2:
            type: oauth2
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/dialog
                    tokenUrl: /oauth/token
    schemas:
...
        ReadAccess:
            type: object
            properties:
                user_id:
                    type: string
                    example: "111458345663774056386"
                topic_id:
                    type: string
                    format: uuid
                last_access:
                    type: string
                    format: date-time
        WriteAccess:
            type: object
            properties:
                user_id:
                    type: string
                    example: "111458345663774056386"
                topic_id:
                    type: string
                    format: uuid
                last_access:
                    type: string
                    format: date-time
paths:
...
    auth/google/callback:
        get:
            summary: OAuth2.0 API endpoint for Google, using PassportJS
            description: Redirects user to Google OAuth2.0 screen, and redirects to eventual page depending on success or failure
            tags:
                - Authentication
            responses:
                "200": # status code
                    description: Redirects to landing page /
                "401": # status code
                    description: Redirects to failure API - /auth/failure
...
    auth/read:
        get:
            summary: Get specific user's read accesses (currently logged in user)
            description: retrieves and returns currently logged in user's read accesses for topics
            tags:
                - Authentication
            responses:
                "200": # status code
                    description: Object containing response details, and array of read accesses
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: boolean
                                        example: false
                                    message:
                                        type: string
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/ReadAccess"
                "403": #status code
                    description: Error, not authorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: boolean
                                    message:
                                        type: string
                                        example: Error, no user logged in
                                    data:
                                        type: array
                                        example: []
                "500": #status code
                    description: Internal Error, item not found or problem finding
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: boolean
                                    message:
                                        type: string
                                        example: Error in retrieving read access
                                    data:
                                        type: array
                                        example: []
...
    auth/write:
        get:
            summary: Get specific user's write accesses (currently logged in user)
            description: retrieves and returns currently logged in user's write accesses for topics
            tags:
                - Authentication
            responses:
                "200": # status code
                    description: Object containing response details, and array of write accesses
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: boolean
                                        example: false
                                    message:
                                        type: string
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/WriteAccess"
                "403": #status code
                    description: Error, not authorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: boolean
                                    message:
                                        type: string
                                        example: Error, no user logged in
                                    data:
                                        type: array
                                        example: []
                "500": #status code
                    description: Internal Error, item not found or problem finding
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: boolean
                                    message:
                                        type: string
                                        example: Error in retrieving write access
                                    data:
                                        type: array
                                        example: []
...